Move resourcemanager/controller/utils.EvalGenericPredicate to resourcemanager/predicate
